// client/lootlion-web/src/app/api/generated/model/wishlistItemDto.ts
/**
 * Lootlion API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WishlistItemDto { 
    id?: string;
    householdId?: string;
    requestedByUserId?: string;
    title?: string | null;
    description?: string | null;
    suggestedCoinCost?: number | null;
    status?: string | null;
    approvedRewardId?: string | null;
    createdUtc?: string;
}